add `unstable_loadStyles` function to foundations
