+++ .changeset/mean-terms-search.md
@@ -1,0 +1,5 @@
+---
+"@stratakit/foundations": patch
+---
+
+Turned `react` and `react-dom` into _optional_ peer dependencies.
