+++ .changeset/ten-kids-cross.md
@@ -1,0 +1,5 @@
+---
+"@stratakit/foundations": patch
+---
+
+Added a new `unstable_loadStyles` function for loading all foundations CSS without using React.
