+++ packages/foundations/package.json
@@ -70,6 +70,14 @@
 		"react": ">=18.0.0",
 		"react-dom": ">=18.0.0"
 	},
+	"peerDependenciesMeta": {
+		"react": {
+			"optional": true
+		},
+		"react-dom": {
+			"optional": true
+		}
+	},
 	"wireit": {
 		"build": {
 			"dependencies": [
