+++ packages/foundations/src/index.ts
@@ -6,3 +6,4 @@
 
 export { Icon } from "./Icon.js";
 export { Root } from "./Root.js";
+export { unstable_loadStyles } from "./styles.js";
