+++ packages/foundations/src/~styles.css.ts
@@ -3,5 +3,5 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 
-import rawStyles from "./styles.css?inline";
+import rawStyles from "./~styles.css?inline";
 export default rawStyles;
